Refactor resource section to enhance carousel functionality and improve component structure
- Renamed `ResourceSection` to `ResourceCarrouselSection` for clarity and consistency.
- Introduced `PageBuilderCarrouselResource` component to encapsulate carousel logic and improve modularity.
- Removed the deprecated `resources.component.tsx` file to streamline the codebase.
- Updated `TestingResources` to utilize the new `ResourceCarrouselSection` for better resource display.
